Provide a way for consumers to not treat . in a markup file as special
